+++ app/platforms/cosmere/character.rb
@@ -15,6 +15,8 @@ module Cosmere
     attribute :investiture, :integer, default: 2
     attribute :selected_skills, array: true, default: {} # { 'acrobatics' => 2, 'arcana' => 1, 'crafting' => 3, 'id' => 1 }
     attribute :additional_skills, array: true, default: {} # { 'id' => { 'name' => '', 'ability' => 'str' } }
+    attribute :expertises, array: true, default: { 'weapon' => [], 'armor' => [], 'culture' => [] }
+    attribute :custom_expertises, array: true, default: [] # [{ 'name' => '', 'desc' => '' }]
     # для левелинга
     attribute :attribute_points, :integer
     attribute :skill_points, :integer