+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Cosmere dice rolls
 - leveling for Cosmere characters
 - selecting heroic path for Cosmere characters
+- selecting expertises for Cosmere characters
 
 ## [0.4.29] - 2026-04-21
 ### Added